Split the reducer search in two parts: criteria and result
