+++ src/example/middleware/middleware-search.js
@@ -1,5 +1,4 @@
 import {builderAdvancedSearchMiddleware} from '../../middleware/middleware-search';
 import {searchAction, otherSearchAction} from '../actions/search-actions';
-
 export const middlewareAdvancedSearch = builderAdvancedSearchMiddleware(searchAction, 'advanced_search', ['SELECTED_FACETS' ]);
 export const middlewareOtherSearch = builderAdvancedSearchMiddleware(otherSearchAction, 'other_search');
