+++ src/middleware/middleware-search.js
@@ -1,15 +1,15 @@
 const SEARCH_UPDATE_QUERY = 'SEARCH_UPDATE_QUERY';
 const DEFAULT_LISTEN_TYPES = ['QUERY'/*, SELECTEDFACETS*/ ]
 import {capitalize, toUpper} from 'lodash/string';
-import {isArray, isFunction,isString} from 'lodash/lang';
+import {isArray, isFunction,isString, isObject} from 'lodash/lang';
 const SEARCH_MIDDLEWARE_BUILDER = 'SEARCH_MIDDLEWARE_BUILDER';
 const STRING_EMPTY = '';
 
 
 
 const _validateActionBuilderParams = ({searchActionAdvancedSearch, name, listenedTypes}) => {
-    if(!isFunction(searchActionAdvancedSearch)) {
-        throw new Error(`${SEARCH_MIDDLEWARE_BUILDER}: the searchActionAdvancedSearch parameter should be a function.`);
+    if(!isObject(searchActionAdvancedSearch)) {
+        throw new Error(`${SEARCH_MIDDLEWARE_BUILDER}: the searchActionAdvancedSearch parameter should be a object.`);
     }
     if(!isString(name) || STRING_EMPTY === name) {
         throw new Error(`${SEARCH_MIDDLEWARE_BUILDER}: the name parameter should be a string`);
