+++ src/reducer/__tests__/reducer-search.js
@@ -1,0 +1,9 @@
+import {createForm, destroyForm, SYNC_FORMS_ENTITY, toggleFormEditing} from '../../actions/form';
+import formReducer from '../form';
+import isArray from 'lodash/isArray';
+
+describe('The search reducer', () => {
+    describe('when receiving a CREATE_FORM action', () => {
+        // TO DO 
+    });
+});
