+++ src/reducer/index.js
@@ -1,3 +1,7 @@
-import searchReducer from './reducer-search'
+import {unitResultsSearchReducerBuilder} from './reducer-result-search';
+import {unitCriteriaSearchReducerBuilder} from './reducer-criteria-search';
+import {combineReducers} from 'redux';
 
-export const search = searchReducer; 
+export const unitSearchReducerBuilder = (name) => {
+  return combineReducers({results : unitResultsSearchReducerBuilder(name), criteria : unitCriteriaSearchReducerBuilder(name)})
+}
